+++ ManuscriptPipeline/Models/CrossrefApplyResult.vb
@@ -1,0 +1,15 @@
+Namespace Models
+
+    Public Class CrossrefApplyResult
+
+        Public Property AuthorLibraryChanged As Boolean = False
+
+        Public Property AuthorsAddedToManuscript As Integer = 0
+
+        Public Property AuthorsCreatedInLibrary As Integer = 0
+
+        Public Property AffiliationsCreatedInLibrary As Integer = 0
+
+    End Class
+
+End Namespace